fix(reconciliation): close manual submission hardening gaps
Why:
- the final hardening pass surfaced module-size, lint, and standards-guard issues inside the manual submission workflow
- the repo's full quality-gates runner also needed an absolute coverage config for subprocess-safe full-suite verification

What:
- split manual submission row validation into a dedicated capability-local module and rename the internal submitted evidence row contract to avoid the balance-evidence ownership guard collision
- tighten submit status assembly so the manual submission package stays pylint-clean without changing the external workflow contract
- make the quality-gates pytest environment pin an absolute coverage config path and add regression coverage for that verification path

Checks:
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run ruff check src/tallylot/application/checkpoints/balance_submission tests/unit/application/checkpoints/balance_submission tests/contract/test_standards_guards.py tools/run_quality_gates.py tests/unit/test_quality_gates.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run mypy src/tallylot/application/checkpoints/balance_submission tools/run_quality_gates.py tests/unit/test_quality_gates.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pyright src/tallylot/application/checkpoints/balance_submission
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pylint src/tallylot/application/checkpoints/balance_submission/validation.py src/tallylot/application/checkpoints/balance_submission/row_validation.py src/tallylot/application/checkpoints/balance_submission/submit.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pytest -o addopts='' tests/unit/application/checkpoints/balance_submission tests/contract/test_standards_guards.py::test_balance_evidence_has_single_production_owner tests/contract/test_standards_guards.py::test_src_does_not_accumulate_flat_same_prefix_clusters tests/unit/test_quality_gates.py -q
- COVERAGE_FILE=/tmp/tallylot-quality-gates-coverage-final UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run python -m tools.run_quality_gates --full-tests
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run tallylot checkpoint submit-balances --source manual-balance-smoke
- python3 .agents/skills/balance-submission-operations/scripts/balance_submission_operations.py run --source manual-balance-smoke
